pyproject.toml
[project]
name = "design-accessibility-mcp"
version = "0.85.2"
description = "Design accessibility analysis MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.1",
    "numpy>=2.2.6",
    "openai>=1.93.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[tool.uv]
package = true

[project.scripts]
context-augment = "context_augment.main:main"
design-accessibility-test = "design_accessibility_mcp.main:main"

